:memo: Add loginRegister Package
ViewModel, Activity, fragment 부분에 loginRegister 패키지를 추가했습니다.

related to: #49
