+++ farming_customer/app/src/main/AndroidManifest.xml
@@ -78,7 +78,7 @@
             android:name=".activity.myPageManagement.MyPageManagementActivity"
             android:exported="false" />
         <activity
-            android:name=".activity.LoginActivity"
+            android:name=".activity.loginRegister.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
