+++ farming_customer/app/src/main/java/kr/co/lion/farming_customer/fragment/loginRegister/FindPasswordFragment.kt
@@ -1,4 +1,4 @@
-package kr.co.lion.farming_customer.fragment
+package kr.co.lion.farming_customer.fragment.loginRegister
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,9 +8,9 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import kr.co.lion.farming_customer.LoginFragmentName
 import kr.co.lion.farming_customer.R
-import kr.co.lion.farming_customer.activity.LoginActivity
+import kr.co.lion.farming_customer.activity.loginRegister.LoginActivity
 import kr.co.lion.farming_customer.databinding.FragmentFindPasswordBinding
-import kr.co.lion.farming_customer.viewmodel.FindPwViewModel
+import kr.co.lion.farming_customer.viewmodel.loginRegister.FindPwViewModel
 
 class FindPasswordFragment : Fragment() {
