+++ farming_customer/app/src/main/java/kr/co/lion/farming_customer/fragment/loginRegister/FindPwDone2Fragment.kt
@@ -1,4 +1,4 @@
-package kr.co.lion.farming_customer.fragment
+package kr.co.lion.farming_customer.fragment.loginRegister
 
 import android.os.Bundle
 import android.text.SpannableString
@@ -12,10 +12,8 @@ import androidx.core.content.ContextCompat
 import androidx.databinding.DataBindingUtil
 import kr.co.lion.farming_customer.LoginFragmentName
 import kr.co.lion.farming_customer.R
-import kr.co.lion.farming_customer.activity.LoginActivity
+import kr.co.lion.farming_customer.activity.loginRegister.LoginActivity
 import kr.co.lion.farming_customer.databinding.FragmentFindPwDone2Binding
-import kr.co.lion.farming_customer.databinding.FragmentFindPwDoneBinding
-import kr.co.lion.farming_customer.viewmodel.FindPwDoneViewModel
 
 class FindPwDone2Fragment : Fragment() {
