+++ farming_customer/app/src/main/java/kr/co/lion/farming_customer/fragment/loginRegister/LoginFragment.kt
@@ -1,4 +1,4 @@
-package kr.co.lion.farming_customer.fragment
+package kr.co.lion.farming_customer.fragment.loginRegister
 
 import android.content.Intent
 import android.os.Bundle
@@ -9,10 +9,10 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import kr.co.lion.farming_customer.LoginFragmentName
 import kr.co.lion.farming_customer.R
-import kr.co.lion.farming_customer.activity.LoginActivity
+import kr.co.lion.farming_customer.activity.loginRegister.LoginActivity
 import kr.co.lion.farming_customer.activity.MainActivity
 import kr.co.lion.farming_customer.databinding.FragmentLoginBinding
-import kr.co.lion.farming_customer.viewmodel.LoginViewModel
+import kr.co.lion.farming_customer.viewmodel.loginRegister.LoginViewModel
 
 class LoginFragment : Fragment() {
