+++ farming_customer/app/src/main/java/kr/co/lion/farming_customer/fragment/loginRegister/Register2Fragment.kt
@@ -1,6 +1,5 @@
-package kr.co.lion.farming_customer.fragment
+package kr.co.lion.farming_customer.fragment.loginRegister
 
-import android.content.Intent
 import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
@@ -8,13 +7,12 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.activity.result.ActivityResultLauncher
 import androidx.databinding.DataBindingUtil
 import kr.co.lion.farming_customer.LoginFragmentName
 import kr.co.lion.farming_customer.R
-import kr.co.lion.farming_customer.activity.LoginActivity
+import kr.co.lion.farming_customer.activity.loginRegister.LoginActivity
 import kr.co.lion.farming_customer.databinding.FragmentRegister2Binding
-import kr.co.lion.farming_customer.viewmodel.Register2ViewModel
+import kr.co.lion.farming_customer.viewmodel.loginRegister.Register2ViewModel
 
 class Register2Fragment : Fragment() {
