--- farming_customer/app/src/main/java/kr/co/lion/farming_customer/viewmodel/LoginViewModel.kt
@@ -1,4 +1,0 @@
-package kr.co.lion.farming_customer.viewmodel
-
-class LoginViewModel {
-}
