+++ farming_customer/app/src/main/java/kr/co/lion/farming_customer/viewmodel/loginRegister/Register3ViewModel.kt
@@ -1,0 +1,4 @@
+package kr.co.lion.farming_customer.viewmodel.loginRegister
+
+class Register3ViewModel {
+}
