+++ farming_customer/app/src/main/res/layout/fragment_find_id.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="findIdViewModel"
-            type="kr.co.lion.farming_customer.viewmodel.FindIdViewModel" />
+            type="kr.co.lion.farming_customer.viewmodel.loginRegister.FindIdViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
