+++ farming_customer/app/src/main/res/layout/fragment_find_id_done.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="findIdDoneViewModel"
-            type="kr.co.lion.farming_customer.viewmodel.FindIdDoneViewModel" />
+            type="kr.co.lion.farming_customer.viewmodel.loginRegister.FindIdDoneViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
