+++ farming_customer/app/src/main/res/layout/fragment_find_password.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="findPwViewModel"
-            type="kr.co.lion.farming_customer.viewmodel.FindPwViewModel" />
+            type="kr.co.lion.farming_customer.viewmodel.loginRegister.FindPwViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
