+++ farming_customer/app/src/main/res/layout/fragment_find_pw_done.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="findPwDoneViewModel"
-            type="kr.co.lion.farming_customer.viewmodel.FindPwDoneViewModel" />
+            type="kr.co.lion.farming_customer.viewmodel.loginRegister.FindPwDoneViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
