+++ farming_customer/app/src/main/res/layout/fragment_register.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="registerViewModel"
-            type="kr.co.lion.farming_customer.viewmodel.RegisterViewModel" />
+            type="kr.co.lion.farming_customer.viewmodel.loginRegister.RegisterViewModel" />
     </data>
 
     <LinearLayout
