+++ farming_customer/app/src/main/res/layout/fragment_register2.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="register2ViewModel"
-            type="kr.co.lion.farming_customer.viewmodel.Register2ViewModel" />
+            type="kr.co.lion.farming_customer.viewmodel.loginRegister.Register2ViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
